Finished basic auth0 integration
