+++ .env.example
@@ -1,0 +1,2 @@
+AUTH0_DOMAIN=domain
+AUTH0_CLIENT_ID=id
