+++ src/Components/index.js
@@ -1,0 +1,15 @@
+import AuthenticationButton from "./AuthenticationButtons/AuthenticationButton";
+import LoginButton from "./AuthenticationButtons/LoginButton";
+import LogoutButton from "./AuthenticationButtons/LogoutButton";
+import SignUpButton from "./AuthenticationButtons/SignUpButton";
+import Dropdown from "./Dropdown";
+import Loading from "./Loading";
+
+export {
+  AuthenticationButton,
+  LoginButton,
+  LogoutButton,
+  Dropdown,
+  SignUpButton,
+  Loading,
+};
