+++ src/index.js
@@ -1,3 +1,5 @@
+import Auth0ProviderWithHistory from "./auth/Auth0ProviderWithHistory";
+import { BrowserRouter } from "react-router-dom";
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
@@ -5,9 +7,11 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+  <BrowserRouter>
+    <Auth0ProviderWithHistory>
+      <App />
+    </Auth0ProviderWithHistory>
+  </BrowserRouter>,
   document.getElementById("root")
 );
