+++ src/pages/Home.js
@@ -1,0 +1,13 @@
+import React, { Fragment } from "react";
+import { LoginButton, SignUpButton } from "../components/index";
+
+const Home = () => {
+  return (
+    <Fragment>
+      <LoginButton />
+      <SignUpButton />
+    </Fragment>
+  );
+};
+
+export default Home;
