+++ src/pages/index.js
@@ -1,0 +1,4 @@
+import Home from "./Home";
+import Profile from "./Profile";
+
+export { Home, Profile };
